requirements.txt
django==3.0.5
social-auth-app-django==3.1.0
six==1.14.0
pillow==7.1.1
psycopg2==2.8.5
django-mptt
loremipsum
django-taggit
django-environ
django-modeltranslation